requirements.txt
Flask>=2.0
Flask-CORS>=4.0
reportlab>=4.0
requests>=2.28.0
python-dotenv>=0.20.0
